Address review comments

+++ sql/src/main/java/com/vmware/ddlog/DDlogJooqHelper.java
@@ -2,6 +2,7 @@ package com.vmware.ddlog;
 
 import com.google.common.collect.ImmutableList;
 import com.vmware.ddlog.ir.DDlogIndexDeclaration;
+import com.vmware.ddlog.util.sql.CreateIndexParser;
 import com.vmware.ddlog.util.sql.ParsedCreateIndex;
 import ddlogapi.DDlogAPI;
 import ddlogapi.DDlogException;
@@ -18,25 +19,38 @@ import static com.vmware.ddlog.DDlogJooqProvider.matchExpressionFromWhere;
 
 /**
  * Helper class to DDlogJooqProvider that contains logic for looking up DDlog indexes and auxiliary functions
- * related to WHERE clauses. *
+ * related to WHERE clauses.
+ *
+ * Callers of the class should use addIndex to add indexes, and then call seal() when all indexes have been added.
+ * The class cannot be used until seal() is called.
  */
 public class DDlogJooqHelper {
     private final Map<String, List<Field<?>>> tablesToFields;
-    private final Map<String, List<ParsedCreateIndex>> tablesToIndexesMap;
+    private final Map<String, List<ParsedCreateIndex>> tablesToIndexesMap = new HashMap<>();
     private final Map<String, List<Field<?>>> indexToFieldsMap = new HashMap<>();
     private final DDlogAPI dDlogAPI;
+    private boolean sealed = false;
 
     public DDlogJooqHelper(Map<String, List<Field<?>>> tablesToFields,
-                           Map<String, List<ParsedCreateIndex>> tablesToIndexesMap,
                            DDlogAPI ddlogAPI) {
         this.tablesToFields = tablesToFields;
-        this.tablesToIndexesMap = tablesToIndexesMap;
         this.dDlogAPI = ddlogAPI;
+    }
 
+    public void addIndex(String sqlString) {
+        if (sealed) {
+            throw new DDlogJooqProvider.DDlogJooqProviderException("Cannot add index to sealed JooqHelper");
+        }
+        ParsedCreateIndex parsedIndex = CreateIndexParser.parse(sqlString);
+        this.tablesToIndexesMap.computeIfAbsent(
+                parsedIndex.getTableName().toUpperCase(Locale.ROOT), k -> new ArrayList<>()).add(parsedIndex);
+    }
+
+    public void seal() {
         // Now construct list of fields for each index
         for (ParsedCreateIndex parsedIndex :
-                tablesToIndexesMap.values().stream().flatMap(List::stream).collect(Collectors.toList())) {
-            List<Field<?>> tableFields = tablesToFields.get(parsedIndex.getTableName().toUpperCase(Locale.ROOT));
+                this.tablesToIndexesMap.values().stream().flatMap(List::stream).collect(Collectors.toList())) {
+            List<Field<?>> tableFields = this.tablesToFields.get(parsedIndex.getTableName().toUpperCase(Locale.ROOT));
             List<Field<?>> indexFields = new ArrayList<>();
             for (String indexColumn : parsedIndex.getColumns()) {
                 for (Field<?> f : tableFields) {
@@ -47,6 +61,7 @@ public class DDlogJooqHelper {
             }
             indexToFieldsMap.put(parsedIndex.getIndexName(), indexFields);
         }
+        sealed = true;
     }
 
     // Check for case-insensitive string equality
@@ -63,7 +78,11 @@ public class DDlogJooqHelper {
     }
 
     public boolean sameFields(SqlNodeList list, String tableName) {
-        String[] tableFields = tablesToFields.get(tableName).stream().map(Field::getName).toArray(String[]::new);
+        if (!sealed) {
+            throw new DDlogJooqProvider.DDlogJooqProviderException("Tried to use unsealed JooqHelper");
+        }
+
+        String[] tableFields = this.tablesToFields.get(tableName).stream().map(Field::getName).toArray(String[]::new);
         String[] queryFields =
                 list.getList().stream().map(x -> {
                     // Just in case the identifier is compound, we get the last name in the names list,
@@ -74,19 +93,17 @@ public class DDlogJooqHelper {
         return compareStringArrays(tableFields, queryFields);
     }
 
-    public String[] getIdsFromWhere(SqlBasicCall where) {
-        WhereClauseGetIdentifiers getIdentifiers = new WhereClauseGetIdentifiers();
-        return where.accept(getIdentifiers).toArray(new String[0]);
-
-    }
-
     public List<DDlogRecord> getEntriesByIndex(String tableName,
                                                String[] ids,
                                                SqlBasicCall where, DDlogJooqProvider.QueryContext context)
             throws DDlogException, SQLException {
+        if (!sealed) {
+            throw new DDlogJooqProvider.DDlogJooqProviderException("Tried to use unsealed JooqHelper");
+        }
+
         final List<DDlogRecord> clonedResults = new ArrayList<>();
 
-        List<ParsedCreateIndex> indexes = tablesToIndexesMap.get(tableName);
+        List<ParsedCreateIndex> indexes = this.tablesToIndexesMap.get(tableName);
         if (indexes == null) {
             throw new SQLException(String.format("Cannot delete from table %s: no appropriate primary key" +
                     "and no indexes", tableName));
@@ -116,6 +133,11 @@ public class DDlogJooqHelper {
         );
 
         return clonedResults;
+    }
+
+    public String[] getIdsFromWhere(SqlBasicCall where) {
+        WhereClauseGetIdentifiers getIdentifiers = new WhereClauseGetIdentifiers();
+        return where.accept(getIdentifiers).toArray(new String[0]);
     }
 
     /**
